requirements.txt
aenum>=2.2.3
aiohttp>=3.9.5
aiosignal>=1.3.1
amqp>=2.6.1
annotated-types>=0.6.0
anyio>=4.3.0
appdirs>=1.4.4
argh>=0.26.2
asgiref>=3.8.1
astroid>=2.12.12
asttokens>=2.0.8
async-timeout>=4.0.2
attrs>=23.2.0
backcall>=0.2.0
backports.functools-lru-cache>=1.6.1
backports.shutil-get-terminal-size>=1.0.0
beautifulsoup4>=4.9.1
billiard>=3.6.3.0
boto3>=1.14.10
botocore>=1.17.63
certifi>=2022.9.24
cffi>=1.15.1
charset-normalizer>=2.0.12
configparser>=5.3.0
contextlib2>=0.6.0.post1
convertdate>=2.4.0
cryptography>=38.0.1
cssselect>=1.1.0
DAWG-Python>=0.7.2
decorator>=5.1.1
Deprecated>=1.2.13
diff-match-patch>=20200713
dill>=0.3.5.1
distlib>=0.3.6
distro>=1.9.0
Django>=4.2.14
django-cors-headers>=3.14.0
django-debug-toolbar>=4.4.6
django-leaflet>=0.28.3
django-mathfilters>=1.0.0
django-modeltranslation>=0.18.2
django-phonenumber-field>=7.3.0
django-redis>=4.12.1
django-reversion>=5.0.12
django-reversion-compare>=0.15.0
django-rosetta>=0.9.8
django-subdomains2>=4.1.2
django-timezone-field>=5.0
django-user-agents>=0.4.0
djantimat @ git+https://github.com/alexander-koval/djantimat.git@7a8fba088519cc53349ad182426ac04c1f0fe0b2
dnspython>=2.2.1
docopt>=0.6.2
docutils>=0.15.2
dodgy>=0.2.1
EditorConfig>=0.12.3
exceptiongroup>=1.2.1
executing>=1.2.0
filelock>=3.8.0
firebirdsql>=1.2.2
frozenlist>=1.4.1
future>=0.18.2
geoip2>=4.8.0
gevent>=22.10.1
greenlet>=1.1.3.post0
h11>=0.14.0
h3>=3.7.6
hiredis>=1.0.1
holidays>=0.10.2
httpcore>=1.0.5
httpx>=0.27.0
idna>=3.4
ipaddr>=2.2.0
ipaddress>=1.0.23
ipwhois>=1.1.0
ipython>=8.5.0
isort>=4.3.21
jedi>=0.18.1
Jinja2>=3.1.2
jmespath>=0.10.0
jsbeautifier>=1.14.11
jsonpatch>=1.33
jsonpointer>=2.4
kombu>=4.6.11
korean-lunar-calendar>=0.3.1
LatLon>=1.0.2
lazy-object-proxy>=1.7.1
lxml>=4.9.1
markdown-it-py>=3.0.0
MarkupSafe>=2.1.3
matplotlib-inline>=0.1.6
matrix-client>=0.4.0
maxminddb>=2.6.1
mccabe>=0.7.0
mdurl>=0.1.2
memory-profiler>=0.57.0
memray>=1.9.1
modernize>=0.7
msgpack>=1.0.5
multidict>=6.0.5
mysql-connector-python>=8.0.31
Naked>=0.1.31
neo4j-driver>=4.3.6
neobolt>=1.7.17
neomodel>=4.0.8
networkx>=2.8.7
numpy>=1.23.4
olefile>=0.46
openai>=1.23.6
packaging>=21.3
paho-mqtt>=1.5.0
parso>=0.8.3
pathlib2>=2.3.5
pathtools>=0.1.2
pep8-naming>=0.4.1
pexpect>=4.8.0
phonenumberslite>=8.13.27
pickleshare>=0.7.5
Pillow>=10.4.0
pip-upgrade-outdated>=1.5
platformdirs>=2.5.2
polib>=1.2.0
prompt-toolkit>=3.0.31
protobuf>=3.19.4
psutil>=5.9.3
psycopg2-binary>=2.8.5
ptyprocess>=0.7.0
pure-eval>=0.2.2
pycodestyle>=2.9.1
pycparser>=2.21
pycurl>=7.45.3
pydantic>=2.7.1
pydantic_core>=2.18.2
pydocstyle>=3.0.0
pydub>=0.24.1
pyflakes>=2.1.1
pygeoip>=0.3.2
pyGeoTile>=1.0.6
Pygments>=2.13.0
pylint>=2.16.0.dev0
pylint-celery>=0.3
pylint-django>=2.5.3
pylint-plugin-utils>=0.7
PyMeeus>=0.5.12
pymorphy2>=0.8
pymorphy2-dicts>=2.4.393442.3710985
pymssql>=2.2.7
pyOpenSSL>=22.1.0
pyparsing>=3.0.9
pyproj>=3.4.0
pyquery>=1.4.1
pyrfc6266>=1.0.2
PySocks>=1.7.1
pytesseract>=0.3.4
python-dateutil>=2.8.2
pytils>=0.4.1
pytz>=2023.3.post1
PyYAML>=5.3.1
pyzmq>=25.0.0b1
redis>=5.0.1
requests>=2.27.1
requirements-detector>=0.7
rich>=13.5.2
s3transfer>=0.3.7
scandir>=1.10.0
scipy>=1.9.3
setoptconf>=0.3.0
Shapely>=1.7.1
shellescape>=3.8.1
singledispatch>=3.4.0.3
six>=1.16.0
sniffio>=1.3.1
snowballstemmer>=2.0.0
sorl-thumbnail>=12.9.0
soupsieve>=2.3.2.post1
sqlparse>=0.4.3
stack-data>=0.5.1
timezonefinder>=6.5.2
toml>=0.10.1
tomli>=2.0.1
tomlkit>=0.11.5
tqdm>=4.66.2
traitlets>=5.5.0
typing>=3.7.4.1
typing_extensions>=4.11.0
tzdata>=2023.3
u-msgpack-python>=2.6.0
ua-parser>=0.16.1
ujson>=5.10.0
urllib3>=1.25.11
user-agents>=2.2.0
uwsgi>=2.0.26
uwsgitop>=0.11
vine>=1.3.0
virtualenv>=20.0.25
watchdog>=0.10.2
wcwidth>=0.2.6
websocket-client>=0.57.0
wrapt>=1.14.1
yarl>=1.9.4
zerorpc @ git+https://github.com/alexander-koval/zerorpc-python.git@bca0ab28ffa62a98329374426b0113722eced55c
zope.event>=4.5.0
zope.interface>=6.2
gtfs-realtime-bindings>=1.0.0
